# public/openapi.yaml
openapi: 3.0.3
info:
  title: Finanzas App API
  description: |
    API REST de la aplicación de finanzas personales (Next.js App Router + Prisma + PostgreSQL).

    ## Autenticación
    La mayoría de los endpoints requieren un **access token JWT** en el header
    `Authorization: Bearer <token>`. El token se obtiene en `/auth/login` o `/auth/register`.
    El **refresh token** se entrega como cookie `HttpOnly` (`refresh_token`) y se usa en `/auth/refresh`.

    ## Convenciones
    - Montos monetarios: número decimal (la API los serializa a veces como string desde PostgreSQL).
    - Fechas: ISO 8601 (`YYYY-MM-DDTHH:mm:ss.sssZ`).
    - Errores de validación → `422` con `{ error, details }`.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Listar movimientos (paginado)
      parameters:
        - { name: page,       in: query, schema: { type: integer, default: 1 } }
        - { name: limit,      in: query, schema: { type: integer, default: 30, maximum: 100 } }
        - { name: from,       in: query, schema: { type: string, format: date-time } }
        - { name: to,         in: query, schema: { type: string, format: date-time } }
        - { name: accountId,  in: query, schema: { type: string } }
        - { name: type,       in: query, schema: { type: string, enum: [INCOME, EXPENSE, TRANSFER] } }
        - { name: categoryId, in: query, schema: { type: string } }
      responses:
        '200':
          description: Página de movimientos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Transaction' } }
                  meta:
                    type: object
                    properties:
                      total: { type: integer }
                      page:  { type: integer }
                      limit: { type: integer }
                      pages: { type: integer }
    post:
      tags: [Transactions]
      summary: Crear movimiento
      description: |
        Movimiento de cuenta única (INCOME/EXPENSE) o **doble partida** (TRANSFER / pago de TDC)
        cuando se incluye `destinationAccountId`. Ajusta saldos de forma atómica.
        - `notes='MSI'` → la deuda se rastrea aparte y NO mueve el saldo.
        - `notes='TDC_PAYMENT'` → se excluye del resumen mensual.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreate' }
      responses:
        '201': { description: Movimiento creado, content: { application/json: { schema: { $ref: '#/components/schemas/Transaction' } } } }
        '404': { description: Cuenta (o destino) no encontrada }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /transactions/summary:
    get:
      tags: [Transactions]
      summary: Totales del mes (Dashboard)
      description: |
        Suma de ingresos y gastos del rango, **excluyendo** traspasos (`type=TRANSFER`)
        y pagos de tarjeta (`notes='TDC_PAYMENT'`) para evitar doble contabilidad.
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to,   in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Totales
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalIncome:  { type: number, example: 25000 }
                  totalExpense: { type: number, example: 18430.5 }
                  net:          { type: number, example: 6569.5 }

  # ────────────────────── CATEGORIES ──────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token JWT obtenido en /auth/login o /auth/register'
  responses:
...
    ValidationError:
      description: Datos inválidos
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string, example: Datos inválidos }
              details: { type: object, additionalProperties: true }
  schemas:
...
    Transaction:
      type: object
      properties:
        id:                { type: string }
        accountId:         { type: string }
        categoryId:        { type: string, nullable: true }
        type:              { type: string, enum: [INCOME, EXPENSE, TRANSFER] }
        amount:            { type: number }
        description:       { type: string }
        date:              { type: string, format: date-time }
        notes:             { type: string, nullable: true, description: "'MSI', 'TDC_PAYMENT', 'RECUR:<freq>' o libre" }
        transferAccountId: { type: string, nullable: true, description: 'Cuenta destino (transferencias/pago TDC)' }
        syncStatus:        { type: string, enum: [SYNCED, PENDING, CONFLICT] }
        account:           { type: object, nullable: true }
        category:          { type: object, nullable: true }
    TransactionCreate:
      type: object
      required: [accountId, type, amount, description, date]
      properties:
        accountId:            { type: string }
        categoryId:           { type: string }
        type:                 { type: string, enum: [INCOME, EXPENSE, TRANSFER] }
        amount:               { type: number, minimum: 0.01 }
        description:          { type: string, maxLength: 200 }
        date:                 { type: string, format: date-time }
        notes:                { type: string, maxLength: 500 }
        clientId:             { type: string, description: 'Para deduplicación offline' }
        destinationAccountId: { type: string, description: 'Cuenta destino (TRANSFER / pago de TDC)' }
...